# eee_portal_backend/openapi.yml
openapi: 3.0.3
info:
  title: EEE Portal Backend API
  description: |
    API documentation for the EEE Portal Backend with complete security specifications.
    
    ## Authentication
    - **Bearer Token**: JWT token required for authenticated endpoints
    - **User Roles**: admin, user
    
    ## Security Levels
    - 🔓 Public: No authentication required
    - 🔐 User: Any logged-in user
    - 🔑 Admin: Admin role required
  version: 1.0.0
  contact:
    name: EEE Portal Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.eeportal.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
          example: 1
        question:
          type: string
          example: "What is the capital of France?"
        question_type:
          type: string
          example: "mcq"
        options:
          type: array
          items:
            type: string
          example: ["Paris", "London", "Berlin", "Madrid"]
        correct_option:
          type: integer
          example: 0
        is_active:
          type: boolean
          example: true
        question_section:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
paths:
...
  /questions/active/:
    get:
      summary: Get active questions
      description: 🔑 Admin only - Retrieve all active questions for assessment
      tags:
        - Questions Service
      operationId: get_active_questions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active questions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
  /questions/{id}/toggle-active/:
    patch:
      summary: Toggle question active status
      description: 🔑 Admin only - Soft enable/disable question without deletion
      tags:
        - Questions Service
      operationId: toggle_question_active
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Question ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Question status toggled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  is_active:
                    type: boolean
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
        '404':
          description: Question not found
  /questions/create/:
    post:
      summary: Create question
      description: 🔑 Admin only - Create new question(s)
      tags:
        - Questions Service
      operationId: create_question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Question'
                - type: array
                  items:
                    $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Question(s) created successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Question'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Question'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
  /questions/update/:
    put:
      summary: Update question
      description: 🔑 Admin only - Update existing question(s)
      tags:
        - Questions Service
      operationId: update_question
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - allOf:
                    - $ref: '#/components/schemas/Question'
                    - type: object
                      required:
                        - id
                - type: array
                  items:
                    allOf:
                      - $ref: '#/components/schemas/Question'
                      - type: object
                        required:
                          - id
      responses:
        '200':
          description: Question(s) updated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Question'
                  - type: array
                    items:
                      $ref: '#/components/schemas/Question'
        '400':
          description: Invalid input or missing ID
        '401':
          description: Unauthorized - Token missing or invalid
        '403':
          description: Forbidden - Admin access required
        '404':
          description: Question not found
...